Reformat the schema dump
trimmed_definition now trims all whitespace on all lines, down to the least
number of spaces. It's kind of a brute force to remove the spaces that seem to
be creeping in.

also change the templating of to_schema

if the sql definition is:
SELECT
 1
FROM
 dual
scenic's to_schema would interpolate it like:
  create_view view_name, sql_definition: <<-SQL
____--SELECT
--  1
--FROM
--  dual
  end
SQL

... where the ____ are the indenting from the heredoc and the
-- is the indenting from the indent(2). This means the SELECT
gets indented twice - once as part of the heredoc, and then again
with the definition.indent - whereas the rest of the sql only
gets indented once with the indent(2).

These two changes are related because when we were calling `.strip`
on the sql it was removing all the leading whitespace from it, placing
it 6 characters in (4 from the HEREDOC, 2 from the indent), but
every other line in the sql was getting indented 2 more additional
whitespaces from the indent(2). So we make the trimmed_sql more
deterministic by constantly removing whitespace from the left of all
lines and moving the interpolation and indentation of the sql
definition to the start of the line and calling indent(6) so all
lines get indented equally. 6 spaces matches where scenic was placing
the SELECT before.

Note we are trimming each line by the least amount of whitespace, so
```
    SELECT
     1
    FROM
       dual
```
becomes
```
SELECT
 1
FROM
   dual
```

and

```
    SELECT
    1
   FROM
    dual
```
becomes
```
 SELECT
 1
FROM
 dual
```

This allows us to preserve users formatting while keeping the schema rb
consistent

Co-authored-by: Davin Lagerroos <dlagerro@umn.edu>
Co-authored-by: Eric Eklund <eeklund@umn.edu>