Generate a consistent schema.rb file
If we load an existing schema that contains `create_view` statements and then
dump the schema (e.g. if a test script calls `db:migrate` before running tests)
the schema file changes even though the database objects are the same. This can
also happen when tests use `ActiveRecord::Migration.maintain_test_schema!`.
This may be all or part of #23.

For example, say we start with a schema.rb that contains a scenic view:
```
  create_view "my_view", sql_defintion: SQL
      SELECT
        1 AS value
      FROM
        dual
  SQL
```
The sql definition, while it has multiple lines, is a single string, i.e.:
`"        SELECT\n        1 AS value\n      FROM\n        dual\n"`

schema:load creates the view using `trimmed_definition` which removes
leading whitespace but preserves the leading whitespace on other lines, so
the view gets built in oracle as:
```
SELECT
        1 AS value
      FROM
        dual
```

If we dump the schema, say after a migration, the view appears as:
```
  create_view "my_view", sql_definition: <<-SQL
      SELECT
          1 AS value
        FROM
          dual
  SQL
```
* the first line, `SELECT`, matches the original formatting
* all other lines are indented 2 more spaces than before

The issue relates to scenic's `to_schema` method. It uses a heredoc as a
template for the create_view statement. Inside that heredoc it calls
`indent(2)` on the view definition. This, combined with the way the view
is built in oracle, is leading to spaces accumulating in the schema.rb
file.

Here is scenic's `to_schema`:
```
    def to_schema
      materialized_option = materialized ? "materialized: true, " : ""

      <<-DEFINITION
  create_view #{UnaffixedName.for(name).inspect}, #{materialized_option}sql_definition: <<-\SQL
    #{escaped_definition.indent(2)}
  SQL
      DEFINITION
    end
```
The `indent(2)` adds 2 leading spaces to each line of the string. In the first
line, i.e. `SELECT`, these spaces (plus the extra spaces from
`escaped_definition` getting interpolated starting four spaces in, under th "e"
of `create_view`) are removed by `trimmed_definition` before the view is built,
so the there is no net effect from multiple loads and dumps.

The extra spacing on the other lines is not removed by `trimmed_definition`.
`trimmed_definition` uses `strip`, which only removes space from the start of
the string and end of the string, not all the lines within the string. As a
result, when the view is built in oracle, the extra `indent(2)` spaces on those
lines become part of the view text in `user_views`. When the database schema is
dumped, the view, with these extra spaces from the previous load, is rendered
by `to_schema`, which adds 2 more spaces. If that schema is loaded and the
database gets dumped later on, the schema.rb file will show two more spaces for
these lines, etc.

To address this we update `trimmed_definition` to use [`strip_heredoc`](https://api.rubyonrails.org/classes/String.html#method-i-strip_heredoc).
to remove a consistent amount of leading space from all lines of the view
definition and rstrip rather than strip to remove trailing whitespace and
newlines from the definition.

We also add a View class to this adapter so we can have a custom `to_schema`
method. This allows us to work around formatting issues caused by most lines
being indented only by `indent(2)`, while the first line gets indented by both
`indent(2)` and the position of the interpolation inside the DEFINITION heredoc.

We added a test to simulate this behavior. One implementation detail for this
test - we added a helper method to the spec_helper to dump the schema. This
makes sure we get a fresh dumper each time. The dumper seems to cache the state
of the database from when it was initialized so reusing it does not detect
schema changes.

Co-authored-by: Davin Lagerroos <dlagerro@umn.edu>
Co-authored-by: Eric Eklund <eeklund@umn.edu>