WIP extract helper method to dump the schema
This makes the test read better and most importantly gets us a fresh
schema dumper each time so the dumper is aware the current state of
the database each time. Before when we were using the same dumper,
we'd get the same schema dump even if the view definition had
changed.

Co-authored-by: Davin Lagerroos <dlagerro@umn.edu>
Co-authored-by: Eric Eklund <eeklund@umn.edu>